adding culling

+++ som.py
@@ -157,7 +157,7 @@ class BatchSOM(SOM):
 		freq_data.index_add_(0, min_idx, torch.ones(x.shape[0]).cuda())
 
 		# Store the one-hot grid of occupied nodes
-		self.grid_used += (freq_data != 0).view(self.rows, self.cols)
+		self.grid_used += (freq_data != 0).view(self.rows, self.cols).long()
 		print self.grid_used
 
 		# Compute aggregate data values for each neighborhood
